Updated events structure
